Start outlining reader, add string functions
