+++ Cargo.toml
@@ -2,5 +2,5 @@
 
 resolver = "2"
 
-members = ["store"]
+members = [ "reader","store"]
